requirements.txt
crewai
langchain
langchain-community
langchain-groq
python-dotenv
